+++ routes/calendar.js
@@ -13,6 +13,88 @@ var CalEvent = require('../models/calendar').CalEvent;
 var Mongoose = require("mongoose");
 require('date-format-lite');
 
+//TODO: all calendar get calls should go through a single function
+
+function makeNewCalendar(user, callback) {
+   var userCalendar = {};
+   Calendar.create({
+     name: 'My Calendar',
+     userId: user._id,
+     eventTypes: [{eventTypeId: 0, name: 'Holiday'},{eventTypeId: 1, name: 'Meeting'},{eventTypeId: 2,name: 'Work'},{eventTypeId:3, name:'Appointment'},{eventTypeId: 4, name: 'Birthday'}],
+     people: [{
+       userId: user._id,
+       permission: 'edit'
+     }]
+   }, function(err, calendar){
+     if (err){
+       console.log('Cal DB create error: ', err);
+       //res.status(500).send({error: true, message: 'Calendar Database Error - ' + err.message});
+       return callback( err, null);
+     }
+
+     //--write calendar id back to user
+     userCalendar = calendar;
+     console.log("makeNewCalendar:",calendar);
+     if(user.calendars) {
+      //   user.calendars.push({calendarId: calendar._id});
+      console.log('did this cal add...from push');
+      console.log(userCalendar);
+      console.log(userCalendar._id);
+         User.update({_id: user.id},{$push: {
+           calendars: {calendarId: userCalendar._id}
+          }}, function(err,userCalendar){
+            if(err){
+              console.log(err)
+            }
+          }); //TODO: return function that return callback with calendar
+     } else {
+       console.log("did this cal add...from set");
+       console.log(userCalendar);
+       console.log(userCalendar._id);
+      //   user.calendars = [{calendarId: calendar._id}];
+         User.update({_id: user.id},{$addToSet: {
+           calendars: {calendarId: userCalendar._id}
+          }}, function(err,userCalendar){
+            if(err){
+              console.log(err)
+            }
+          });
+     }
+   })
+}
+
+//-- do not return entire calendar, no events
+function getUserCalendar(user, callback) {
+   var userCalendar = {};
+   //-- user must always have a calendar[0]
+   if(!user.calendars || !user.calendars[0]) {
+      makeNewCalendar(user, function(err, calendar) {
+         if(err) {
+            console.log("db error: could not make new calendar: ",err);
+            return callback( err, null);
+         } else {
+            Calendar.findOne({_id: calendar._id}).select({events:0}).exec( function(err, calendar){
+               if(err){
+                  console.log('DB error - calendar not found: ', err);
+                  return callback( err, null);
+               }
+
+               callback(null, calendar);
+            });
+         }
+      });
+   } else {
+      Calendar.findOne({_id: user.calendars[0].calendarId}).select({events:0}).exec(function(err, calendar){
+         if(err){
+            console.log('DB error - calendar not found: ', err);
+            return callback( err, null);
+         }
+         callback(null, calendar);
+      });
+
+   }
+}
+
 router.post('/all', function(req,res,next){
     console.log(req.body);
     User.aggregate([
@@ -255,4 +337,4 @@ router.post('/editName', function(req,res,next){
 });
 
 
-module.exports = router;
+module.exports = {router, getUserCalendar, makeNewCalendar};
